requirements.txt
APScheduler==3.10.4
httpx==0.27.0
python-dotenv==1.0.1
Quart==0.19.5
supabase==2.4.5
requests==2.32.3
markdown2==2.4.13
postgrest==0.16.4
psycopg2-binary==2.9.9